Change rule on else statement of game function
The problem inside game() is when roundResult is 'Ti\!', computerScore increments by 1. Mainly because it lacks conditional statement on else. It should be false when You win! and 'Tie\!' strings are searched. Change the conditional statement if\else to ternary operator and declare two variables outside for loop to store the results in searching if You win! and 'True\!' is in roundResult. Second argument checks if 'You win\!' and 'Tie\!' is not inside search roundResult.